+++ src/main/java/backend/jangbogoProject/commodity/CommodityRepository.java
@@ -2,7 +2,7 @@ package backend.jangbogoProject.commodity;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommodityRepository extends JpaRepository<Commodity, Long> {
+public interface CommodityRepository extends JpaRepository<Commodity, Integer> {
 
 
 }
